+++ packages/node/src/integrations/http/SentryHttpInstrumentationBeforeOtel.ts
@@ -4,7 +4,7 @@ import { flush, logger, vercelWaitUntil } from '@sentry/core';
 import type * as http from 'node:http';
 import type * as https from 'node:https';
 import { DEBUG_BUILD } from '../../debug-build';
-import { stealthWrap } from './shared';
+import { stealthWrap } from './utils';
 
 type Http = typeof http;
 type Https = typeof https;
